Logger: route traces - got 95% of the way there
Required remapping a bunch of SD card pins. Only things that I
couldn't manage to route onto a 4cm by 6cm single sided board was the
current sense line, and a couple of ground pour islands. It should be
doable with maybe one or two jumpers. I want to try rotating the
microcontroller 90 degrees CCW, that should make the current sense
line easier, as well as some of the SD card stuff.