Radio board: Generate netlist and do preliminary layout.
Noticed that one footprint was misassigned - the MCP2562 was listed as
a 14 pin package when it should have been 8 pins

Done preliminary sizing. I believe that if the board is packed very
tightly, it can be made as a 6x7cm 6/6 2-layer board. Might be that
I'm overoptimistic, the preliminary layout doesn't even fit in that
space, but heck, I'm hopeful.