Logger: Shuffle layout a bit, reduce board size to 4.5x4.5cm
The traces in this commit are ugly, they need to be cleaned up, but
all nodes that need to be connected are. Need to space out some
components so that we don't want to die when assembling the thing.

This commit puts both SD cards together. IMHO, it's much more
aesthetic than it was before. And, things are connected