Logger: First Review-Ready version
Things not covered in this review: ground trace widths. They're too
small. I know. That'll take some finnickying.

Things that should be reviewed: Layout, sizing, is there enough space
to solder everything. Is doing a 1-sided board a terrible idea? Are
the traces nice (they're not)/are they workeable (maybe?)

This commit christens this board RSRLog - The rocket-fast System for
rocket logging. Name subject to change, I'm not 100% attached to it.