Sensor mini: Add analog stuff
Adds one thermistor and the tank pressure sensor to sensor_mini.
Required making the board a quarter inch wider. This commit updates the
PCB and the schematic.

Note that while all the traces are there (DRC mostly passes (IMU is too
small for 6mil DRC), and everything is connected), it isn't done. The
traces are not great, much improvement possible. Also, it's likely that
we'll be swapping the integrated IMU and gyro for just a discrete gyro,
since it was a bitch to work with on the big sensor board.