Remove limit switch traces from comparators to MCU
There was an issue with gyoza V2 in which when the microcontroller was
depowered it loaded the limit switch traces through the protection
diodes on the microcontroller's pins. This caused mosfets to conduct
when they shouldn't, the linear actuator to move erroneously, and much
gnashing of teeth.

This change is tested and known to work, the gyoza 2 that we're
currently using for all our testing and such has had these two traces
removed, and it works fine. Analog pot feedback is still connected,
which is why we have position reporting. The analog signal doesn't
_seem_ to be loaded by the protection diodes, because in closed state it
has a voltage of 0.5, below the conduction point of your average diode.