Replace valve open DAC with resistive divider
We need the vent valve to open once the board receives battery power,
potentially before the MCU wakes up and is able to initialize the DACs.
So, we use the OG DAC, the resistive divider. It's referenced to 5VBAT,
so it will work regardless of whether bus power is applied. It should
generate 0.5V, which is what we're currently using on gyoza 2, which
works.