Connect linear actuator pot VREF to 5VBAT, not 5V
There was a bug in the last schematic in which the linear actuator
feedback pot was referenced to the microcontroller/bus 5V rail, so when
the microcontroller was powered we had no way of checking the linear
actuator position. Now we reference it to 5VBAT, so whenever the valve
battery is connected, the linear actuator feedback is also present.

In order to remain compatible with the bodged version of gyoza2 that
we've been testing on, the potentiometer positive reference is on pin 6
of the linear actuator connector, not pin 1 as it was on the two
previous versions of the board.

This commit reworks a bunch of the 5VBAT routing, because it was daisy
chaining hard in a few places. It also increases the width of the 5VBAT
trace form 6mil to 2.5mm (9.84mil), because I like making power traces
fatter.