Add initial tracing
Definitely not super happy with the first round but I wanted something
to get out for review so we can get this started. I mentioned in slack
we needed an actual part number but it won't affect the number of pads
on Q1 so it should be easy to replace even if the footprint changes.